PBinary implementation
